Improve compatability with extension modules
Python extension modules such as Numpy are currently causing
issues when the FMU is imported by a tool that has not loaded
libpython3.x.so previously. This resulted in FMPy being able to
load the FMU whereas the compliance checker and into-cps would
fail.

fmi2Functions:
- added dlopen call during initialization of Python interpreter.
 It uses the following call to determine the name of the shared object:
 >>> sysconfig.get_config_var("INSTSONAME")
 libpython3.7m.so.1.0
 It remains unclear how robust this apporach is and if a more suitable
 alternative exists.

It still remains unclear why modules loading pyfmu.so do not
automatically import the libpython. However, it seems the issue is
encountered frequenly when combining embedding and import of
Numpy:
https://stackoverflow.com/questions/11842920/undefined-symbol
-pyexc-importerror-when-embedding-python-in-c