Fix unload of shared libraries in tests.
Some test cases either directly loads pyfmu.so or do so implictly thorugh
FMPy. However, they would not unload the library. On Windows this causes
issues with clean up since the DLL would still be loaded causing an access error
when trying to delete the extracted FMU.

Solution is simply to do the unloading